feat(ReactApplication): Adds Application base class
There are two goals of this base class:
1. Reduce boiler plate in the generated solutions for react-native-windows
2. Manage things like URL activations, lifecycle state subscriptions, etc. from within react-native-windows.

This is only the first step, we will likely also need to create something like a ReactApplicationDelegate, similar to the ReactActivityDelegate to separate these initialization behaviors from a concrete base class as all apps will not be able to use the ReactApplication base class.
