+++ local-cli/generator-windows/templates/src/App.xaml
@@ -1,8 +1,8 @@
-﻿<Application
+﻿<rn:ReactApplication
     x:Class="<%= ns %>.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:<%= ns %>"
+    xmlns:rn="using:ReactNative"
     RequestedTheme="Light">
 
-</Application>
+</rn:ReactApplication>
